move tree handlers to its own file
