+++ src/mocks/handlers/document/index.ts
@@ -1,4 +1,5 @@
 import { handlers as recycleBinHandlers } from './recycle-bin.handlers.js';
+import { handlers as treeHandlers } from './tree.handlers.js';
 import { handlers as documentHandlers } from './document.handlers.js';
 
-export const handlers = [...recycleBinHandlers, ...documentHandlers];
+export const handlers = [...recycleBinHandlers, ...treeHandlers, ...documentHandlers];
